+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Breaking
+
+### Changed
+
+- Organized default metrics
+
+### Added
+
 ## [0.16.0] - 2026-08-24
 
 This release marks our first release as a Prometheus subproject.